requirements.txt
pandas>=2.2
numpy>=1.26
openpyxl>=3.1
PyYAML>=6.0
neo4j>=5.20
scikit-learn>=1.5
